Update from task 49c4b7de-53e7-433d-ae51-34089ca734d2

+++ app/database/models/admin_action.py
@@ -1,7 +1,7 @@
 """Admin action audit log model."""
 
 from datetime import datetime
-from sqlalchemy import Integer, String, DateTime, func, ForeignKey, Text, JSON
+from sqlalchemy import BigInteger, String, DateTime, func, ForeignKey, Text, JSON
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 from app.database.session import Base
 
@@ -11,13 +11,13 @@ class AdminAction(Base):
 
     __tablename__ = "admin_actions"
 
-    id: Mapped[int] = mapped_column(Integer, primary_key=True)
+    id: Mapped[int] = mapped_column(BigInteger, primary_key=True)
     admin_id: Mapped[int] = mapped_column(
         ForeignKey("users.id", ondelete="CASCADE"), nullable=False, index=True
     )
     action: Mapped[str] = mapped_column(String(100), nullable=False)
     target_type: Mapped[str] = mapped_column(String(50), nullable=False)  # e.g., "order", "config"
-    target_id: Mapped[int | None] = mapped_column(Integer, nullable=True)
+    target_id: Mapped[int | None] = mapped_column(BigInteger, nullable=True)
     action_metadata: Mapped[dict | None] = mapped_column(JSON, nullable=True)
     created_at: Mapped[datetime] = mapped_column(
         DateTime(timezone=True), server_default=func.now(), nullable=False
